# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateStoreRequest:
    properties:
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
...
  controllers.StoresListResponse:
    properties:
      pagination:
        $ref: '#/definitions/utils.PaginationResponse'
      stores:
        items:
          $ref: '#/definitions/models.StoreResponse'
        type: array
    type: object
...
  controllers.UpdateStoreRequest:
    properties:
      code:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
...
  models.StoreResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: 192.168.31.136:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive user management backend service with JWT authentication
    and role-based access control
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Livotech Backend Service API
  version: "1.0"
paths:
...
  /api/stores:
    get:
      consumes:
      - application/json
      description: Get list of all stores with optional search by Code or Name (logged-in
        users only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by Code or Name (partial match)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.StoresListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get all stores
      tags:
      - stores
    post:
      consumes:
      - application/json
      description: Create a new store (logged-in users only)
      parameters:
      - description: Create Store Request
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create new store
      tags:
      - stores
  /api/stores/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a store (logged-in users only)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Remove store
      tags:
      - stores
    get:
      consumes:
      - application/json
      description: Get specific store by ID (logged-in users only)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StoreResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get store by ID
      tags:
      - stores
    put:
      consumes:
      - application/json
      description: Update store information (logged-in users only)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Store Request
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update store
      tags:
      - stores
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abcdef12345'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"